Refactoring of commands module and Initial Merkle Tree.
