+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod model;
 pub mod storage;
+pub mod validation;
