+++ src/model/commands/accounts.rs
@@ -1,0 +1,213 @@
+/// The purpose of add signatory command is to add an identifier to the account. Such
+/// identifier is a public key of another device or a public key of another user.
+#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct AddSignatory {
+    pub account_id: String,
+    pub public_key: [u8; 32],
+}
+
+/// # Example
+/// ```
+/// use iroha::model::commands::accounts::AddSignatory;
+///
+/// let command_payload = AddSignatory {
+///     account_id: "account@domain".to_string(),
+///     public_key: [63; 32],
+/// };
+/// let result: Vec<u8> = command_payload.into();
+/// ```
+impl std::convert::From<AddSignatory> for Vec<u8> {
+    fn from(command_payload: AddSignatory) -> Self {
+        bincode::serialize(&command_payload).expect("Failed to serialize payload.")
+    }
+}
+
+/// # Example
+/// ```
+/// # use iroha::model::commands::accounts::AddSignatory;
+/// #
+/// # let command_payload = AddSignatory {
+/// #     account_id: "account@domain".to_string(),
+/// #     public_key: [63; 32],
+/// # };
+/// # let result: Vec<u8> = command_payload.into();
+/// let command_payload: AddSignatory = result.into();
+/// ```
+impl std::convert::From<Vec<u8>> for AddSignatory {
+    fn from(command_payload: Vec<u8>) -> Self {
+        bincode::deserialize(&command_payload).expect("Failed to deserialize payload.")
+    }
+}
+
+#[test]
+fn add_signatory_command_serialization_and_deserialization() {
+    let expected = AddSignatory {
+        account_id: "account@domain".to_string(),
+        public_key: [63; 32],
+    };
+    let actual: AddSignatory =
+        bincode::deserialize(&bincode::serialize(&expected).unwrap()[..]).unwrap();
+    assert_eq!(expected, actual);
+}
+
+/// The purpose of append role command is to promote an account to some created role in the
+/// system, where a role is a set of permissions account has to perform an action (command or
+/// query).
+#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct AppendRole {
+    pub account_id: String,
+    pub role_name: String,
+}
+
+/// # Example
+/// ```
+/// use iroha::model::commands::accounts::AppendRole;
+///
+/// let command_payload = AppendRole {
+///     account_id: "account@domain".to_string(),
+///     role_name: "role".to_string(),
+/// };
+/// let result: Vec<u8> = command_payload.into();
+/// ```
+impl std::convert::From<AppendRole> for Vec<u8> {
+    fn from(command_payload: AppendRole) -> Self {
+        bincode::serialize(&command_payload).expect("Failed to serialize payload.")
+    }
+}
+
+/// # Example
+/// ```
+/// # use iroha::model::commands::accounts::AppendRole;
+/// #
+/// # let command_payload = AppendRole {
+/// #     account_id: "account@domain".to_string(),
+/// #     role_name: "role".to_string(),
+/// # };
+/// # let result: Vec<u8> = command_payload.into();
+/// let command_payload: AppendRole  = result.into();
+/// ```
+impl std::convert::From<Vec<u8>> for AppendRole {
+    fn from(command_payload: Vec<u8>) -> Self {
+        bincode::deserialize(&command_payload).expect("Failed to deserialize payload.")
+    }
+}
+
+#[test]
+fn append_role_command_serialization_and_deserialization() {
+    let expected = AppendRole {
+        account_id: "account@domain".to_string(),
+        role_name: "role".to_string(),
+    };
+    let actual: AppendRole =
+        bincode::deserialize(&bincode::serialize(&expected).unwrap()[..]).unwrap();
+    assert_eq!(expected, actual);
+}
+
+/// The purpose of create account command is to make entity in the system, capable of sending
+/// transactions or queries, storing signatories, personal data and identifiers.
+#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct CreateAccount {
+    pub account_name: String,
+    pub domain_id: String,
+    pub public_key: [u8; 32],
+}
+
+/// # Example
+/// ```
+/// use iroha::model::commands::accounts::CreateAccount;
+///
+/// let command_payload = CreateAccount {
+///     account_name: "account".to_string(),
+///     domain_id: "domain".to_string(),
+///     public_key: [63; 32],
+/// };
+/// let result: Vec<u8> = command_payload.into();
+/// ```
+impl std::convert::From<CreateAccount> for Vec<u8> {
+    fn from(command_payload: CreateAccount) -> Self {
+        bincode::serialize(&command_payload).expect("Failed to serialize payload.")
+    }
+}
+
+/// # Example
+/// ```
+/// # use iroha::model::commands::accounts::CreateAccount;
+/// #
+/// # let command_payload = CreateAccount {
+/// #     account_name: "account".to_string(),
+/// #     domain_id: "domain".to_string(),
+/// #     public_key: [63; 32],
+/// # };
+/// # let result: Vec<u8> = command_payload.into();
+/// let command_payload: CreateAccount  = result.into();
+/// ```
+impl std::convert::From<Vec<u8>> for CreateAccount {
+    fn from(command_payload: Vec<u8>) -> Self {
+        bincode::deserialize(&command_payload).expect("Failed to deserialize payload.")
+    }
+}
+
+#[test]
+fn create_account_command_serialization_and_deserialization() {
+    let expected = CreateAccount {
+        account_name: "account".to_string(),
+        domain_id: "domain".to_string(),
+        public_key: [63; 32],
+    };
+    let actual: CreateAccount =
+        bincode::deserialize(&bincode::serialize(&expected).unwrap()[..]).unwrap();
+    assert_eq!(expected, actual);
+}
+
+/// The purpose of create role command is to create a new role in the system from the set of
+/// permissions. Combining different permissions into roles, maintainers of Iroha peer network
+/// can create customized security model.
+#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct CreateRole {
+    pub role_name: String,
+    pub permissions: Vec<String>,
+}
+
+/// # Example
+/// ```
+/// use iroha::model::commands::accounts::CreateRole;
+///
+/// let command_payload = CreateRole {
+///     role_name: "user".to_string(),
+///     permissions: Vec::new(),
+/// };
+/// let result: Vec<u8> = command_payload.into();
+/// ```
+impl std::convert::From<CreateRole> for Vec<u8> {
+    fn from(command_payload: CreateRole) -> Self {
+        bincode::serialize(&command_payload).expect("Failed to serialize payload.")
+    }
+}
+
+/// # Example
+/// ```
+/// # use iroha::model::commands::accounts::CreateRole;
+/// #
+/// # let command_payload = CreateRole {
+/// #    role_name: "user".to_string(),
+/// #    permissions: Vec::new(),
+/// # };
+/// # let result: Vec<u8> = command_payload.into();
+/// let command_payload: CreateRole  = result.into();
+/// ```
+impl std::convert::From<Vec<u8>> for CreateRole {
+    fn from(command_payload: Vec<u8>) -> Self {
+        bincode::deserialize(&command_payload).expect("Failed to deserialize payload.")
+    }
+}
+
+#[test]
+fn create_role_command_serialization_and_deserialization() {
+    let expected = CreateRole {
+        role_name: "user".to_string(),
+        permissions: Vec::new(),
+    };
+    let actual: CreateRole =
+        bincode::deserialize(&bincode::serialize(&expected).unwrap()[..]).unwrap();
+    assert_eq!(expected, actual);
+}
