+++ src/model/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod accounts;
+pub mod assets;
+pub mod domains;
+pub mod peers;
